Clean code: Move Transfer and TransferType to its own package.
