+++ src/test/java/org/opentripplanner/gtfs/mapping/TransferMapperTest.java
@@ -1,19 +1,10 @@
 package org.opentripplanner.gtfs.mapping;
 
-import org.junit.Test;
 import org.onebusaway.gtfs.model.AgencyAndId;
 import org.onebusaway.gtfs.model.Route;
 import org.onebusaway.gtfs.model.Stop;
 import org.onebusaway.gtfs.model.Transfer;
 import org.onebusaway.gtfs.model.Trip;
-
-import java.util.Collection;
-import java.util.Collections;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
 
 public class TransferMapperTest {
     private static final String FEED_ID = "FEED";
@@ -87,7 +78,7 @@ public class TransferMapperTest {
     /*
     @Test
     public void testMap() throws Exception {
-        org.opentripplanner.model.Transfer result = subject.map(TRANSFER);
+        org.opentripplanner.model.transfer.Transfer result = subject.map(TRANSFER);
 
         assertNotNull(result.getFromRoute());
         assertNotNull(result.getFromTrip());
@@ -104,7 +95,7 @@ public class TransferMapperTest {
     /*
     @Test
     public void testMapWithNulls() throws Exception {
-        org.opentripplanner.model.Transfer result = subject.map(new Transfer());
+        org.opentripplanner.model.transfer.Transfer result = subject.map(new Transfer());
 
         assertNull(result.getFromRoute());
         assertNull(result.getFromTrip());
@@ -122,8 +113,8 @@ public class TransferMapperTest {
     /*
     @Test
     public void testMapCache() throws Exception {
-        org.opentripplanner.model.Transfer result1 = subject.map(TRANSFER);
-        org.opentripplanner.model.Transfer result2 = subject.map(TRANSFER);
+        org.opentripplanner.model.transfer.Transfer result1 = subject.map(TRANSFER);
+        org.opentripplanner.model.transfer.Transfer result2 = subject.map(TRANSFER);
 
         assertTrue(result1 == result2);
     }
